+++ noodles-bam/examples/bam_query.rs
@@ -17,9 +17,9 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
     let region = args.next().expect("missing region").parse()?;
 
     let mut reader = bam::indexed_reader::Builder::default().build_from_path(src)?;
-    let header: sam::Header = reader.read_header()?.parse()?;
+    let header = reader.read_header()?.parse()?;
 
-    let query = reader.query(header.reference_sequences(), &region)?;
+    let query = reader.query(&header, &region)?;
 
     let stdout = io::stdout().lock();
     let mut writer = sam::Writer::new(stdout);